chore: refactor coordination
